// src/api/transaction/enums.ts
export enum OperationTypes {
  CREDIT = 'CREDIT',
  DEBIT = 'DEBIT',
}

export enum TransactionTypes {
  BUY_CRYPTO = 'BUY_CRYPTO',
  CASH_IN = 'CASH_IN',
  CASH_OUT = 'CASH_OUT',
  DEPOSIT_FINANCE = 'DEPOSIT_FINANCE',
  REVERT_BUY = 'REVERT_BUY',
  REVERT_CASH_OUT = 'REVERT_CASH_OUT',
  SELL_CRYPTO = 'SELL_CRYPTO',
  START = 'START',
  TRANSFER_BAAS_WALLET = 'TRANSFER_BAAS_WALLET',
  TRANSFER_IN_P2P_WALLET = 'TRANSFER_IN_P2P_WALLET',
  TRANSFER_OUT_P2P_WALLET = 'TRANSFER_OUT_P2P_WALLET',
  WITHDRAW_FINANCE = 'WITHDRAW_FINANCE',
  MIGRATION_BAAS_DOCK = 'MIGRATION_BAAS_DOCK',
  CASH_IN_PAYMENT_SLIP = 'CASH_IN_PAYMENT_SLIP',
  CASH_OUT_MANUAL_PAYMENT = 'CASH_OUT_MANUAL_PAYMENT',
}

export enum Sort {
  ASC = 'ASC',
  DESC = 'DESC',
}

export enum AccountStatus {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
}

export enum AccountSegmentation {
  NORMAL = 'NORMAL',
}